Aggiuto Maven e portati i sorgenti a UTF-8
