+++ src/main/java/it/neo7bf/Database.java
@@ -1,0 +1,18 @@
+package it.neo7bf;
+
+import java.util.HashMap;
+import java.util.Map;
+
+//per ora senza la gestione delle eccezioni
+
+public class Database {
+	
+	private Map<String,ContoCorrente> conti = new HashMap<String,ContoCorrente>();
+
+	public void aggiungiContoCorrente(String carta, ContoCorrente conto) {
+		conti.put(carta, conto);
+	}
+	public ContoCorrente getContoCorrente(String carta) {
+		return conti.get(carta);
+	}
+}
